Update README.md
This repository was made for my weather app dashboard that shows the current weather, a 5 day forecast, and a list of passed location searches.

I used HTML, CSS, JavaScript, JQuery and a OpenWeatherMap API to make this app.

This app is easy to use. After landing on the landing page, you can search the weather of any given city. The App uses local storage to hold onto your passed searches so you can revisit the app and see where you searched. This is a great feature to have.